add separate interceptor for SKU
